--- src/lib/helpers/GetKey.js
@@ -1,3 +1,0 @@
-export const getKey = (prefix, currentKey, parentKeyPath, marginLeft) => {
-  return `${prefix}_${parentKeyPath}_${currentKey}_${marginLeft}`;
-};
