--- src/lib/helpers/ViewerKey.js
@@ -1,3 +1,0 @@
-export const getViewerKey = (prefix, currentKey, parentKeyPath, marginLeft) => {
-  return `${prefix}_${parentKeyPath}_${currentKey}_${marginLeft}`;
-};
