--- src/lib/helpers/editKey.js
@@ -1,1 +1,0 @@
-export const EDIT_KEY = "__editable_json_editor__";
